auth/requirements.txt
# PATCHED v0.2.3 auth/requirements.txt — widen fastapi version range

fastapi>=0.111,<1.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.1
python-multipart==0.0.6
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
redis==5.0.4
